requirements.txt
flask==3.1.0
apscheduler==3.10.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-docx==1.1.2
pymongo==4.10.1
anthropic==0.42.0
openai==1.59.0
